+++ src/main/java/net/sourceforge/pmd/util/fxdesigner/DesignerStarter.java
@@ -4,12 +4,11 @@
 
 package net.sourceforge.pmd.util.fxdesigner;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
-import javax.swing.JOptionPane;
+import static net.sourceforge.pmd.util.fxdesigner.util.JavaFxUtil.isCompatibleJavaFxVersion;
+import static net.sourceforge.pmd.util.fxdesigner.util.JavaFxUtil.isJavaFxAvailable;
+import static net.sourceforge.pmd.util.fxdesigner.util.JavaFxUtil.setSystemProperties;
 
-import org.apache.commons.lang3.SystemUtils;
+import javax.swing.JOptionPane;
 
 import net.sourceforge.pmd.annotation.InternalApi;
 
@@ -30,18 +29,7 @@ public final class DesignerStarter {
             + " Please install the latest JavaFX on your system and try again." + System.lineSeparator()
             + " See https://gluonhq.com/products/javafx/";
 
-    private static final int MIN_JAVAFX_VERSION_ON_MAC_OSX = 14;
-
     private DesignerStarter() {
-    }
-
-    private static boolean isJavaFxAvailable() {
-        try {
-            DesignerStarter.class.getClassLoader().loadClass("javafx.application.Application");
-            return true;
-        } catch (ClassNotFoundException | LinkageError e) {
-            return false;
-        }
     }
 
     /**
@@ -52,41 +40,6 @@ public final class DesignerStarter {
     public static void main(String[] args) {
         final ExitStatus ret = launchGui(args);
         System.exit(ret.getCode());
-    }
-
-    private static void setSystemProperties() {
-        if (SystemUtils.IS_OS_LINUX) {
-            // On Linux, JavaFX renders text poorly by default. These settings help to alleviate the problems.
-            System.setProperty("prism.text", "t2k");
-            System.setProperty("prism.lcdtext", "true");
-        }
-    }
-    
-    private static boolean isCompatibleJavaFxVersion() {
-        if (SystemUtils.IS_OS_MAC_OSX) {
-            final String javaFxVersion = getJavaFxVersion();
-            if (javaFxVersion != null) {
-                final int major = Integer.parseInt(javaFxVersion.split("\\.")[0]);
-                if (major < MIN_JAVAFX_VERSION_ON_MAC_OSX) {
-                    // Prior to JavaFx 14, text on Mac OSX was garbled and unreadable
-                    return false;
-                }
-            }
-        }
-
-        return true;
-    }
-
-    private static String getJavaFxVersion() {
-        try (InputStream is = DesignerStarter.class.getClassLoader().getResourceAsStream("javafx.properties")) {
-            final Properties javaFxProperties = new Properties();
-            javaFxProperties.load(is);
-            return (String) javaFxProperties.get("javafx.version");
-        } catch (IOException ignored) {
-            // Can't determine the version
-        }
-
-        return null;
     }
 
     @SuppressWarnings("PMD.AvoidCatchingThrowable")